Export code in a detectMissingVendors function of react-dev-utils
